pi/signal_detection/requirements.txt
# pi/signal_detection/requirements.txt
pyserial
websockets
asyncio
